requirements-dev.txt
# ============================================================================
# BLACKBOX5 - DEVELOPMENT DEPENDENCIES
# ============================================================================
# Purpose: Development and testing dependencies for Blackbox5
# Last Updated: 2026-02-10
# ============================================================================

# Include all runtime dependencies
-r requirements.txt

# ============================================================================
# TESTING FRAMEWORK
# ============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0

# Test utilities
hypothesis>=6.92.0
factory-boy>=3.3.0
faker>=22.0.0

# ============================================================================
# CODE QUALITY & LINTING
# ============================================================================
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
pylint>=3.0.0
mypy>=1.8.0

# Type stubs
types-redis>=4.6.0
types-pyyaml>=6.0.0
types-psutil>=5.9.0

# ============================================================================
# DOCUMENTATION
# ============================================================================
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocstrings[python]>=0.24.0

# ============================================================================
# SECURITY SCANNING
# ============================================================================
bandit>=1.7.0
safety>=2.3.0

# ============================================================================
# PROFILING & DEBUGGING
# ============================================================================
memray>=1.11.0
py-spy>=0.3.0
line-profiler>=4.1.0

# ============================================================================
# GITHUB INTEGRATION
# ============================================================================
# For task sync and PR automation
PyYAML>=6.0.1
requests>=2.31.0
markdown>=3.5.0
